Simplify handling of negations and optionals via nested graphs
